+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## 2.3.2
+- Bumped snyk dependency to 1.24.6
+
 ## 2.3.1
 - Added url field to Feedback model
 - Fixed CSS static calls for feedback pages